Remove meta.pipeline, meta.folderId, add meta.detection
The main reason for this is to simplify the metadata stored on
folders/items. Before, the meta.pipeline field was not being used in the
changed steps, and just caused confusion. The meta.folderId field was
replaced with meta.detection, as folderId is a field that already exists
on an item, and meta.detection conveys that the current item contains
detections for the folder which corresponds to the value of meta.detection.