Add volume cleaning step to deploy.yml
This allows the two services that use the pipelines volume to view
the updates from new images. If this step is not performed, updating the
image and rebuilding will not reflect the updates to the directory in
the new image, as any updates will be overwritten by the previously stored
data in that volume.