Add: ONNX Export
The main objective of this commit was to add a 'ONNX export' button to export trained models on the platform.

Export to ONNX function is available on Dive Web & Desktop.
Added a tab to list, delete and export all trained pipelines.
Added the possibility to download single or multiple files from the browser on Dive web
Fixed alignment issue on Jobs list on Dive web